+++ ReleaseNotes.md
@@ -1,7 +1,15 @@
-## Release Notes ##
+# Release Notes
 
-## v. 0.1.0 ##
+## v. 0.2.0
 
-### Enhancements ###
+### Enhancements
+
+* Added support for shortcuts `Ctrl+Alt+<Num_1-5>` - switching between sidebar
+elements `Chat`, `Meet`, `Communities`, `Calendar`, `Activity`.
+
+
+## v. 0.1.0
+
+### Enhancements
 
 * Informing about unread chats/activities detected